Moved linalg tests into appropriately named files
